+++ server/plugin.go
@@ -31,12 +31,12 @@ type Plugin struct {
 
 // CloudClient is the interface for managing cloud installations.
 type CloudClient interface {
-	GetClusters(*cloud.GetClustersRequest) ([]*cloud.Cluster, error)
-
-	CreateInstallation(request *cloud.CreateInstallationRequest) (*cloud.Installation, error)
-	GetInstallation(installationID string, request *cloud.GetInstallationRequest) (*cloud.Installation, error)
-	GetInstallations(*cloud.GetInstallationsRequest) ([]*cloud.Installation, error)
-	UpdateInstallation(installationID string, request *cloud.PatchInstallationRequest) (*cloud.Installation, error)
+	GetClusters(*cloud.GetClustersRequest) ([]*cloud.ClusterDTO, error)
+	CreateInstallation(request *cloud.CreateInstallationRequest) (*cloud.InstallationDTO, error)
+	GetInstallation(installationID string, request *cloud.GetInstallationRequest) (*cloud.InstallationDTO, error)
+	GetInstallationByDNS(DNS string, request *cloud.GetInstallationRequest) (*cloud.InstallationDTO, error)
+	GetInstallations(*cloud.GetInstallationsRequest) ([]*cloud.InstallationDTO, error)
+	UpdateInstallation(installationID string, request *cloud.PatchInstallationRequest) (*cloud.InstallationDTO, error)
 	DeleteInstallation(installationID string) error
 
 	GetClusterInstallations(request *cloud.GetClusterInstallationsRequest) ([]*cloud.ClusterInstallation, error)
